requirements.txt
cassandra-driver==3.26.0
click==8.1.3
Flask==2.2.3
Flask-CQLAlchemy==2.0.0
geomet==0.2.1.post1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
six==1.16.0
Werkzeug==2.2.3
yapf==0.32.0